Move Yaz to /sp/
