+++ configure.py
@@ -204,7 +204,6 @@ code_in_files = {
         os.path.join('payload', 'game', 'system', 'SaveManager.c'),
         os.path.join('payload', 'game', 'system', 'SceneCreatorDynamic.S'),
         os.path.join('payload', 'game', 'system', 'SceneCreatorDynamic.c'),
-        os.path.join('payload', 'game', 'system', 'Yaz.c'),
         os.path.join('payload', 'game', 'ui', 'ControlLoader.S'),
         os.path.join('payload', 'game', 'ui', 'Font.S'),
         os.path.join('payload', 'game', 'ui', 'FontManager.c'),
@@ -299,6 +298,7 @@ code_in_files = {
         os.path.join('payload', 'sp', 'Slab.c'),
         os.path.join('payload', 'sp', 'Storage.c'),
         os.path.join('payload', 'sp', 'Tcp.c'),
+        os.path.join('payload', 'sp', 'Yaz.c'),
         os.path.join('vendor', 'arith64.c'),
         os.path.join('vendor', 'ff', 'diskio.c'),
         os.path.join('vendor', 'ff', 'ff.c'),
